backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML APIs
google-generativeai==0.3.2
elevenlabs==0.2.27
requests==2.31.0

# Video processing
moviepy==1.0.3
ffmpeg-python==0.2.0
Pillow==10.2.0

# Audio processing
pydub==0.25.1

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1

